services/scanners/dns/requirements.txt
starlette
gunicorn
uvicorn
emoji
pytest
requests>=2.18.4
checkdmarc
dkimpy
dnspython
tldextract
PyNaCl
pretend
uvloop
httptools
Pebble==4.6.1
